+++ src/components/settings/Settings.svelte
@@ -67,6 +67,7 @@
         />
     </Link>
     <LanguageChooser />
+
     <Dialog
         button={{
             tip: (l) => l.ui.dialog.settings.button.show,